Moved files around, refactored simluation code
Changed the simluation functionality so that it would append to the
end of the states array rather than having a set length defined
at the start. This makes sense in the context of using the model
as a simulator with a controller. Added inputs parameter to the
update function to allow for variable inputs.